+++ tests/packerlicious/test_post_processor_docker.py
@@ -19,3 +19,13 @@ class TestDockerPushPostProcessor(object):
         b = post_processor.DockerPush()
 
         b.to_dict()
+
+
+class TestDockerImportPostProcessor(object):
+
+    def test_required_fields_missing(self):
+        b = post_processor.DockerSave()
+
+        with pytest.raises(ValueError) as excinfo:
+            b.to_dict()
+        assert 'required' in str(excinfo.value)