+++ CHANGELOG.md
@@ -6,4 +6,5 @@
 * builder/file: add packer file builder support
 * provisioner/DockerImport: add packer docker import provisioner support
 * provisioner/DockerPush: add packer docker push provisioner support
+* provisioner/DockerSave: add packer docker save provisioner support
 * provisioner/file: add packer file provisioner support